LayoutTests/platform/mac/accessibility/visible-character-range-scrolling-expected.txt

This tests that visibleCharacterRange returns expected visible ranges with various scrollTop values.


Testing scrollTop values 0 to 2000.
Range with view 200x500, scrollTop 0: NSRange: {0, 809}
Range with view 200x500, scrollTop 100: NSRange: {179, 779}
Range with view 200x500, scrollTop 200: NSRange: {329, 809}
Range with view 200x500, scrollTop 300: NSRange: {509, 809}
Range with view 200x500, scrollTop 400: NSRange: {659, 809}
Range with view 200x500, scrollTop 500: NSRange: {839, 809}
Range with view 200x500, scrollTop 600: NSRange: {989, 809}
Range with view 200x500, scrollTop 700: NSRange: {1169, 809}
Range with view 200x500, scrollTop 800: NSRange: {1319, 809}
Range with view 200x500, scrollTop 900: NSRange: {1499, 809}
Range with view 200x500, scrollTop 1000: NSRange: {1679, 779}
Range with view 200x500, scrollTop 1100: NSRange: {1829, 809}
Range with view 200x500, scrollTop 1200: NSRange: {2009, 809}
Range with view 200x500, scrollTop 1300: NSRange: {2159, 809}
Range with view 200x500, scrollTop 1400: NSRange: {2339, 809}
Range with view 200x500, scrollTop 1500: NSRange: {2489, 809}
Range with view 200x500, scrollTop 1600: NSRange: {2669, 809}
Range with view 200x500, scrollTop 1700: NSRange: {2819, 809}
Range with view 200x500, scrollTop 1800: NSRange: {2999, 809}
Range with view 200x500, scrollTop 1900: NSRange: {3179, 779}
Range with view 200x500, scrollTop 2000: NSRange: {3329, 809}

PASS successfullyParsed is true

TEST COMPLETE
